Fix a compiler warning in some compiler versions
Fix type conversion warning (unsigned long long to uint64_t) that is treated as an error

+++ src/arch/fpga/fpgainstrumentation.cpp
@@ -2,10 +2,10 @@
 #include "libxdma.h"
 
 unsigned long long int FPGAInstrumentation::getDeviceTime() {
-   unsigned long long time;
+   uint64_t time;
    //TODO error checking
    xdmaGetDeviceTime(&time);
-   return time;
+   return (unsigned long long)time;
 }
 
 unsigned long long FPGAInstrumentation::translateDeviceTime( unsigned long long devTime ) {
